add local alias
