+++ examples/demo-app/package.json
@@ -1,13 +1,11 @@
 {
   "scripts": {
-    "start": "webpack-dev-server --mode development --progress --hot --open",
-    "start:esbuild": "node esbuild.config.mjs --start",
-    "start-prod": "webpack-dev-server -p --progress --hot --open",
-    "start-local-deck": "webpack-dev-server --mode development --env.es6 --env.deck --progress --hot --open",
-    "start-local-deck-src": "webpack-dev-server --mode development --env.es6 --env.deck_src --progress --hot --open",
-    "start-local-loaders-src": "webpack-dev-server --mode development --env.es6 --env.loaders_src --progress --hot --open",
-    "start-local-hubble-src": "webpack-dev-server --mode development --env.es6 --env.hubble_src --progress --hot --open",
-    "start-local-https": "webpack-dev-server --mode development --https --env.es6 --progress --hot --open"
+    "start": "node esbuild.config.mjs --start",
+    "start:local": "NODE_ENV=local node esbuild.config.mjs --start --env.deck",
+    "start:local-deck": "NODE_ENV=local node esbuild.config.mjs --start --env.deck",
+    "start:local-deck-src": "NODE_ENV=local node esbuild.config.mjs --start --env.deck_src",
+    "start:local-loaders-src": "NODE_ENV=local node esbuild.config.mjs --start --env.loaders_src",
+    "start:local-hubble-src": "NODE_ENV=local node esbuild.config.mjs --start --env.hubble_src"
   },
   "dependencies": {
     "@auth0/auth0-spa-js": "^2.1.2",
